+++ lib/src/responses/models/models.dart
@@ -4,3 +4,5 @@ export './ask.dart';
 export './poll.dart';
 export './poll_option.dart';
 export './user.dart';
+export './item.dart';
+export './job.dart';