+++ lib/src/responses/utils/utils.dart
@@ -1,3 +1,6 @@
 class ResponsesUtils {
-  static timeFromJson(int time) => DateTime.fromMicrosecondsSinceEpoch(time);
+  static timeFromJson(int timeInSeconds) {
+    final int timeInMilliSeconds = int.parse('${timeInSeconds}000');
+    return DateTime.fromMillisecondsSinceEpoch(timeInMilliSeconds);
+  }
 }